implement dir for ByObjectRef
